+++ CHANGELOG.md
@@ -1,16 +1,33 @@
 # Changelog
 
-## 3.2.0
+## Unreleased
+
+### Added
+
+* Unlimited (`UINT_64MAX`) signers amount in free license
+
+### Fixed
+
+* Added checking permission for setting `allTasks` in operations: `CreateAMLAlertRequestOp`, `CreateManageLimitsRequestOp`, `CreateIssuanceRequestOp`, `CreatePreIssuanceRequestOp`, `UpdateAssetOp`, `ManageSaleOp`
+
+## 3.3.1
+
+### Fixed
+
+* curl now added in containers
+
+## 3.3.0
 
 ### Added
 
 * Manage account specific rule op frame
 * New CreateAMLAlert error codes
+* Cancel change role reviewable request
 
 ### Fixed
 
 * Create and approve change role request for role that does  not exist
-
+* Removed ability to create sale for reversed pair
 
 ## 3.2.1